+++ Kernel/System/Ticket/Article.pm
@@ -269,7 +269,7 @@ sub ArticleCreate {
     my $RandomString = $Kernel::OM->Get('Kernel::System::Main')->GenerateRandomString(
         Length => 32,
     );
-    my $ArticleInsertFingerprint = $$ . '-' . $RandomString . '-' . ($Param{MessageID} // '');
+    my $ArticleInsertFingerprint = $$ . '-' . $RandomString . '-' . ( $Param{MessageID} // '' );
 
     # get database object
     my $DBObject = $Kernel::OM->Get('Kernel::System::DB');
@@ -469,6 +469,7 @@ sub ArticleCreate {
             TicketID         => $Param{TicketID},
             UserID           => $Param{UserID},
             AutoResponseType => $Param{AutoResponseType},
+            ArticleType      => $Param{ArticleType}
         );
     }
 
@@ -2984,12 +2985,12 @@ sub SendCustomerNotification {
     }
 
     # replace config options
-    $Notification{Body} =~ s{<OTRS_CONFIG_(.+?)>}{$ConfigObject->Get($1)}egx;
+    $Notification{Body}    =~ s{<OTRS_CONFIG_(.+?)>}{$ConfigObject->Get($1)}egx;
     $Notification{Subject} =~ s{<OTRS_CONFIG_(.+?)>}{$ConfigObject->Get($1)}egx;
 
     # cleanup
     $Notification{Subject} =~ s/<OTRS_CONFIG_.+?>/-/gi;
-    $Notification{Body} =~ s/<OTRS_CONFIG_.+?>/-/gi;
+    $Notification{Body}    =~ s/<OTRS_CONFIG_.+?>/-/gi;
 
     # COMPAT
     $Notification{Body} =~ s/<OTRS_TICKET_ID>/$Param{TicketID}/gi;
@@ -3005,7 +3006,7 @@ sub SendCustomerNotification {
     );
     for ( sort keys %Ticket ) {
         if ( defined $Ticket{$_} ) {
-            $Notification{Body} =~ s/<OTRS_TICKET_$_>/$Ticket{$_}/gi;
+            $Notification{Body}    =~ s/<OTRS_TICKET_$_>/$Ticket{$_}/gi;
             $Notification{Subject} =~ s/<OTRS_TICKET_$_>/$Ticket{$_}/gi;
         }
     }
@@ -3015,20 +3016,20 @@ sub SendCustomerNotification {
 
     # cleanup
     $Notification{Subject} =~ s/<OTRS_TICKET_.+?>/-/gi;
-    $Notification{Body} =~ s/<OTRS_TICKET_.+?>/-/gi;
+    $Notification{Body}    =~ s/<OTRS_TICKET_.+?>/-/gi;
 
     # get current user data
     my %CurrentPreferences = $UserObject->GetUserData( UserID => $Param{UserID} );
     for ( sort keys %CurrentPreferences ) {
         if ( $CurrentPreferences{$_} ) {
-            $Notification{Body} =~ s/<OTRS_CURRENT_$_>/$CurrentPreferences{$_}/gi;
+            $Notification{Body}    =~ s/<OTRS_CURRENT_$_>/$CurrentPreferences{$_}/gi;
             $Notification{Subject} =~ s/<OTRS_CURRENT_$_>/$CurrentPreferences{$_}/gi;
         }
     }
 
     # cleanup
     $Notification{Subject} =~ s/<OTRS_CURRENT_.+?>/-/gi;
-    $Notification{Body} =~ s/<OTRS_CURRENT_.+?>/-/gi;
+    $Notification{Body}    =~ s/<OTRS_CURRENT_.+?>/-/gi;
 
     # get owner data
     my %OwnerPreferences = $UserObject->GetUserData(
@@ -3036,14 +3037,14 @@ sub SendCustomerNotification {
     );
     for ( sort keys %OwnerPreferences ) {
         if ( $OwnerPreferences{$_} ) {
-            $Notification{Body} =~ s/<OTRS_OWNER_$_>/$OwnerPreferences{$_}/gi;
+            $Notification{Body}    =~ s/<OTRS_OWNER_$_>/$OwnerPreferences{$_}/gi;
             $Notification{Subject} =~ s/<OTRS_OWNER_$_>/$OwnerPreferences{$_}/gi;
         }
     }
 
     # cleanup
     $Notification{Subject} =~ s/<OTRS_OWNER_.+?>/-/gi;
-    $Notification{Body} =~ s/<OTRS_OWNER_.+?>/-/gi;
+    $Notification{Body}    =~ s/<OTRS_OWNER_.+?>/-/gi;
 
     # get responsible data
     my %ResponsiblePreferences = $UserObject->GetUserData(
@@ -3051,20 +3052,20 @@ sub SendCustomerNotification {
     );
     for ( sort keys %ResponsiblePreferences ) {
         if ( $ResponsiblePreferences{$_} ) {
-            $Notification{Body} =~ s/<OTRS_RESPONSIBLE_$_>/$ResponsiblePreferences{$_}/gi;
+            $Notification{Body}    =~ s/<OTRS_RESPONSIBLE_$_>/$ResponsiblePreferences{$_}/gi;
             $Notification{Subject} =~ s/<OTRS_RESPONSIBLE_$_>/$ResponsiblePreferences{$_}/gi;
         }
     }
 
     # cleanup
     $Notification{Subject} =~ s/<OTRS_RESPONSIBLE_.+?>/-/gi;
-    $Notification{Body} =~ s/<OTRS_RESPONSIBLE_.+?>/-/gi;
+    $Notification{Body}    =~ s/<OTRS_RESPONSIBLE_.+?>/-/gi;
 
     # get ref of email params
     my %GetParam = %{ $Param{CustomerMessageParams} };
     for ( sort keys %GetParam ) {
         if ( $GetParam{$_} ) {
-            $Notification{Body} =~ s/<OTRS_CUSTOMER_DATA_$_>/$GetParam{$_}/gi;
+            $Notification{Body}    =~ s/<OTRS_CUSTOMER_DATA_$_>/$GetParam{$_}/gi;
             $Notification{Subject} =~ s/<OTRS_CUSTOMER_DATA_$_>/$GetParam{$_}/gi;
         }
     }
@@ -3078,14 +3079,14 @@ sub SendCustomerNotification {
         # replace customer stuff with tags
         for ( sort keys %CustomerUser ) {
             if ( $CustomerUser{$_} ) {
-                $Notification{Body} =~ s/<OTRS_CUSTOMER_DATA_$_>/$CustomerUser{$_}/gi;
+                $Notification{Body}    =~ s/<OTRS_CUSTOMER_DATA_$_>/$CustomerUser{$_}/gi;
                 $Notification{Subject} =~ s/<OTRS_CUSTOMER_DATA_$_>/$CustomerUser{$_}/gi;
             }
         }
     }
 
     # cleanup all not needed <OTRS_CUSTOMER_DATA_ tags
-    $Notification{Body} =~ s/<OTRS_CUSTOMER_DATA_.+?>/-/gi;
+    $Notification{Body}    =~ s/<OTRS_CUSTOMER_DATA_.+?>/-/gi;
     $Notification{Subject} =~ s/<OTRS_CUSTOMER_DATA_.+?>/-/gi;
 
     # format body
@@ -3094,7 +3095,7 @@ sub SendCustomerNotification {
     }
     for ( sort keys %Article ) {
         if ( $Article{$_} ) {
-            $Notification{Body} =~ s/<OTRS_CUSTOMER_$_>/$Article{$_}/gi;
+            $Notification{Body}    =~ s/<OTRS_CUSTOMER_$_>/$Article{$_}/gi;
             $Notification{Subject} =~ s/<OTRS_CUSTOMER_$_>/$Article{$_}/gi;
         }
     }
@@ -3106,7 +3107,7 @@ sub SendCustomerNotification {
     );
     if ( $Notification{Subject} =~ /<OTRS_CUSTOMER_SUBJECT\[(.+?)\]>/ ) {
         my $SubjectChar = $1;
-        $Article{Subject} =~ s/^(.{$SubjectChar}).*$/$1 [...]/;
+        $Article{Subject}      =~ s/^(.{$SubjectChar}).*$/$1 [...]/;
         $Notification{Subject} =~ s/<OTRS_CUSTOMER_SUBJECT\[.+?\]>/$Article{Subject}/g;
     }
     $Notification{Subject} = $Self->TicketSubjectBuild(
@@ -3132,7 +3133,7 @@ sub SendCustomerNotification {
     }
 
     # cleanup all not needed <OTRS_CUSTOMER_ tags
-    $Notification{Body} =~ s/<OTRS_CUSTOMER_.+?>/-/gi;
+    $Notification{Body}    =~ s/<OTRS_CUSTOMER_.+?>/-/gi;
     $Notification{Subject} =~ s/<OTRS_CUSTOMER_.+?>/-/gi;
 
     # send notify
@@ -3183,6 +3184,7 @@ send an auto response to a customer via email
             Subject => 'For the message!',
         },
         UserID          => 123,
+        ArticleType     => 'email-internal'  # optional
     );
 
 Events:
@@ -3364,7 +3366,13 @@ sub SendAutoResponse {
             User => $Ticket{CustomerUserID},
         );
 
-        if ( $CustomerUser{UserEmail} && $OrigHeader{From} !~ /\Q$CustomerUser{UserEmail}\E/i ) {
+        $Param{ArticleType} //= '';
+        if (
+            $CustomerUser{UserEmail}
+            && $OrigHeader{From} !~ /\Q$CustomerUser{UserEmail}\E/i
+            && $Param{ArticleType} ne 'email-internal'
+            )
+        {
             $Cc = $CustomerUser{UserEmail};
         }
     }